demo/requirements.txt
diskcache==5.6.3
Jinja2==3.1.6
llama_cpp_python @ git+https://github.com/abetlen/llama-cpp-python.git@c37132bac860fcc333255c36313f89c4f49d4c8d
MarkupSafe==3.0.3
numpy==2.4.0
typing_extensions==4.15.0
